Prepare tcache and PAC for size to grow by PAGE over 2 * NGROUP * PAGE.
To prepare for the upcoming changes where size class grows by PAGE when
larger than 32K, disable the tcache when it is larger than 32K.  To
avoid having too many bins, arena bins still have the same layout.  This
means some extra search is needed for a page-level request that is not
aligned with the orginal size class: it should also search the heap
before the current index since the previous heap might also be able to
have some allocations satisfying it.  For now, a simple one-element
search is added.  It will be changed into a more thorough search with
configurable maximum number of items checked.